requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-bower==5.0.4
django-filter==0.9.2
django-toolbelt==0.0.1
djangorestframework==3.1.1
gunicorn==19.3.0
psycopg2==2.6
six==1.9.0
static3==0.5.1
